Move extensions into util package
