+++ src/main/kotlin/dev/monosoul/jooq/settings/ImageAware.kt
@@ -1,6 +1,6 @@
 package dev.monosoul.jooq.settings
 
-import dev.monosoul.jooq.callWith
+import dev.monosoul.jooq.util.callWith
 import groovy.lang.Closure
 import org.gradle.api.Action
